@@ -23,8 +23,8 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.8.1</version>
         <configuration>
-          <source>9</source>
-          <target>9</target>
+          <source>16</source>
+          <target>16</target>
         </configuration>
       </plugin>
       <plugin>
@@ -107,7 +107,7 @@
       <dependency>
           <groupId>me.clip</groupId>
           <artifactId>placeholderapi</artifactId>
-          <version>2.10.7</version>
+          <version>2.11.4</version>
           <scope>provided</scope>
       </dependency>
       <dependency>
